Create apps for menu, category, item
