--- src/event/serializers/category_item_serializer.py
@@ -1,7 +1,0 @@
-from rest_framework import serializers
-from event.models import CategoryItemAssociation
-
-class CategoryItemAssociationSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = CategoryItemAssociation
-        fields = "__all__"
