--- src/event/serializers/subitem_serializer.py
@@ -1,8 +1,0 @@
-from rest_framework import serializers
-from event.models import SubItem
-
-class SubItemSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = SubItem
-        fields = "__all__"
